insurtech_ocr/requirements.txt
streamlit 
pytesseract 
pdf2image 
pillow